store only once

+++ src/WebService/src/main/java/eu/smartsantander/androidExperimentation/repository/SmartphoneRepository.java
@@ -15,5 +15,5 @@ public interface SmartphoneRepository extends CrudRepository<Smartphone, Long> {
 
     Smartphone findById(int id);
 
-    Smartphone findByPhoneId(int phoneId);
+    Smartphone findByPhoneId(long phoneId);
 }
